updated penalty management to set limit on penalty accrued

+++ backend/src/main/java/com/finapp/backend/loan/entity/LoanInstallment.java
@@ -41,7 +41,7 @@ public class LoanInstallment {
     @Column(nullable = false)
     private LoanPaymentStatus paymentStatus;
 
-    private int daysOverdue;
+    private long daysOverdue;
 
     @Column(nullable = false)
     private LocalDateTime lastUpdated;
